Removing compiler warning in previous commit

+++ src/particle_filter.cpp
@@ -144,8 +144,8 @@ void ParticleFilter::updateWeights(double sensor_range, double std_landmark[],
 		vector<LandmarkObs> trans_observation;
 		for (int j = 0;j < observations.size(); ++j)
 		{
-			float x_map = particles[i].x + (cos(particles[i].theta) * observations[j].x) - (sin(particles[i].theta) * observations[j].y);
-			float y_map = particles[i].y + (sin(particles[i].theta) * observations[j].x) + (cos(particles[i].theta) * observations[j].y);
+			double x_map = particles[i].x + (cos(particles[i].theta) * observations[j].x) - (sin(particles[i].theta) * observations[j].y);
+			double y_map = particles[i].y + (sin(particles[i].theta) * observations[j].x) + (cos(particles[i].theta) * observations[j].y);
 			trans_observation.push_back(LandmarkObs{ x_map, y_map, particles[i].theta });
 		}
